+++ packages/stream_chat_flutter/lib/src/attachment/thumbnail/image_attachment_thumbnail.dart
@@ -3,7 +3,6 @@ import 'dart:io' show File;
 import 'package:flutter/material.dart';
 import 'package:stream_chat_flutter/src/attachment/thumbnail/thumbnail_size_calculator.dart';
 import 'package:stream_chat_flutter/stream_chat_flutter.dart';
-import 'package:stream_core_flutter/stream_core_flutter.dart';
 
 /// {@template imageAttachmentThumbnail}
 /// Widget for building image attachment thumbnail.