+++ packages/stream_chat_localizations/example/lib/add_new_lang.dart
@@ -256,6 +256,9 @@ class NnStreamChatLocalizations extends GlobalStreamChatLocalizations {
   @override
   String get photosLabel => 'Photos';
 
+  @override
+  String get photosAndVideosLabel => 'Photos & Videos';
+
   String _getDay(DateTime dateTime) {
     final now = DateTime.now();
     final today = DateTime(now.year, now.month, now.day);