+++ packages/stream_chat_localizations/lib/src/stream_chat_localizations_de.dart
@@ -229,6 +229,9 @@ class StreamChatLocalizationsDe extends GlobalStreamChatLocalizations {
   @override
   String get photosLabel => 'Fotos';
 
+  @override
+  String get photosAndVideosLabel => 'Fotos & Videos';
+
   String _getDay(DateTime dateTime) {
     final now = DateTime.now();
     final today = DateTime(now.year, now.month, now.day);